pyproject.toml
[project]
name = "rune"
version = "0.4.0"
description = "Agent orchestration for AI coding tools — explore, plan, build, validate."
requires-python = ">=3.12"
dependencies = ["click>=8.0", "ruamel.yaml>=0.18", "rich>=13.0", "jsonschema>=4.0", "jinja2>=3.1"]
license = {text = "MIT"}
readme = "README.md"
keywords = ["ai", "agents", "claude", "coding-assistant", "knowledge-management", "llm-wiki"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.scripts]
rune = "cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cli"]

[tool.hatch.build.targets.wheel.sources]
"src/rune-agency" = "rune-agency"
"platforms" = "platforms"
"profiles.yaml" = "profiles.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.urls]
Homepage = "https://github.com/kvasir-ai/rune"
Repository = "https://github.com/kvasir-ai/rune"
Issues = "https://github.com/kvasir-ai/rune/issues"
